CMake build: changing DLND to SLND in A-compset
As a result of merging in master to this branch,
there was a change in the A compset definition
to use SLND instead of DLND. This mirrors that
change in the Cmake lists.

I verified that the A and S case executables
built via CMake, when copied on top of acme.exe
in a case bld directory, will pass tests.